mini_project2/VS code/requirements.txt
flask
pycryptodome
requests